Feature d v0.08

See merge request 2022-New-Ecosystem-Group/mall-shop!215

+++ service-goods/service-goods-core/src/main/java/com/wanmi/sbc/goods/mini/service/goods/WxGoodsService.java
@@ -436,7 +436,7 @@ public class WxGoodsService {
             BigDecimal price = goodsInfo.getMarketPrice().multiply(BigDecimal.valueOf(100));
             sku.setSalePrice(price);
             sku.setMarketPrice(price);
-            sku.setStockNum(goodsInfo.getStock().intValue());
+            sku.setStockNum(goodsInfo.getStock());
             List<WxAddProductRequest.SkuAttrs> skuAttrsList = new ArrayList<>();
             WxAddProductRequest.SkuAttrs skuAttrs = new WxAddProductRequest.SkuAttrs();
             sku.setSkuCode("樊登读书SkuCode");
